Updates the library, python package, and documentation from 0.4.0 to 0.4.1

[ committed by @ashao ]
[ reviewed by @mellis13, @MattToast ]

+++ examples/serial/c/example_put_get_3D.c
@@ -66,7 +66,7 @@ int main(int argc, char* argv[]) {
   for(size_t i=0; i<dims[0]; i++)
     for(size_t j=0; j<dims[1]; j++)
       for(size_t k=0; k<dims[2]; k++)
-        tensor[i][j][k] = ((double)rand())/RAND_MAX;
+        tensor[i][j][k] = ((double)rand())/(double)RAND_MAX;
 
   char key[] = "3D_tensor_example";
   size_t key_length = strlen(key);
